+++ src/Layers/GroupLayerBehaviour.cs
@@ -14,7 +14,7 @@ namespace DCFApixels.WhimTex
         public GroupCompositing compositing { get => Owner.compositing; set => Owner.compositing = value; }
 
         internal override bool IsGroup => true;
-        internal bool IsPassThrough => compositing == GroupCompositing.PassThrough && swizzle.IsIdentity;
+        internal bool IsPassThrough => Owner.IsPassThrough;
         internal BlendMode EffectiveBlendMode => compositing == GroupCompositing.PassThrough ? BlendMode.Normal : blendMode;
 
         internal override RenderTexture Render(in LayerRenderContext context)